Adds TGMarker class
Users will no longer use marker identifier to manager marker a helper class
that stores the tangram marker states and interfaces with the tangram map
instance.
