+++ platforms/ios/src/TangramMap/TGGeoPolygon.h
@@ -83,6 +83,8 @@ NS_ASSUME_NONNULL_BEGIN
  */
 - (void)removeAll;
 
+- (instancetype)userData;
+
 NS_ASSUME_NONNULL_END
 
 @end
