+++ platforms/ios/src/TangramMap/TGMapViewController+Internal.h
@@ -1,0 +1,16 @@
+//
+//  TGMapViewController+Internal.h
+//  TangramMap
+//
+//  Created by Karim Naaji on 2/17/17.
+//  Copyright (c) 2017 Mapzen. All rights reserved.
+//
+
+#import "TGMapViewController.h"
+#import "tangram.h"
+
+@interface TGMapViewController ()
+
+@property (assign, nonatomic, nullable) Tangram::Map* map;
+
+@end
