+++ platforms/ios/src/TangramMap/TGMarkerPickResult.h
@@ -8,14 +8,9 @@
 
 #import <Foundation/Foundation.h>
 #import "TGGeoPoint.h"
+#import "TGMarker.h"
 
 NS_ASSUME_NONNULL_BEGIN
-
-/**
- A marker identifier for use in the <a href="Classes/TGMapViewController.html#/Marker%20interface">marker interface</a>.
- A marker identifier of 0 is non-valid.
- */
-typedef uint32_t TGMapMarkerId;
 
 /**
  Data structure holding the result of a marker selection that occured on the map view.
@@ -35,13 +30,13 @@ typedef uint32_t TGMapMarkerId;
  @note You shouldn't have to create a `TGMarkerPickResult` yourself, those are returned as a result to
  a selection query on the `TGMapViewController` and initialized by the latter.
  */
-- (instancetype) initWithCoordinates:(TGGeoPoint)coordinates identifier:(TGMapMarkerId)identifier;
+- (instancetype) initWithCoordinates:(TGGeoPoint)coordinates marker:(TGMarker *)marker;
 
 /// The geographic coordinates of the selected label
 @property (readonly, nonatomic) TGGeoPoint coordinates;
 
 /// The identifier of the selected marker
-@property (readonly, nonatomic) TGMapMarkerId identifier;
+@property (readonly, nonatomic) TGMarker* marker;
 
 NS_ASSUME_NONNULL_END
